Crate vgrs

This library lets Rust programs running inside Valgrind make various requests of Valgrind and its tools. For now this only works on Linux, FreeBSD or MacOS, and only on 32- or 64-bit x86, but support for other platforms should be easy.

This crate is sparsely documented. You will probably want to look at the Valgrind user manual and the C headers in /usr/include/valgrind to learn what these requests do.

Some client requests are not implemented, and some of those that are implemented are untested. These requests are from Valgrind 3.8; some of them may not exist on older versions.

Interpreting the ability to trust Valgrind results as an extension of Rust's memory safety guarantee, all of these functions are marked unsafe. Even so, some of them are clearly safe, but are still marked unsafe for consistency.

When not running under Valgrind, these requests do nothing and return a default value (usually zero).

These are #[inline(always)] to match the C versions, which are macros.

Modules

callgrind
Client requests for the Callgrind profiler tool.

drd
Client requests for the DRD thread error detector tool.

helgrind
Client requests for the Helgrind thread error detector tool.

memcheck
Client requests for the Memcheck memory error detector tool.

valgrind
Client requests for the Valgrind core itself.